The simulated 1394 amplifier is compiled as a separate dynamic library
